requirements.txt
pandas
g2p-en == 2.1.0
inflect == 4.1.0
librosa == 0.7.2
matplotlib == 3.2.2
numba == 0.48
numpy == 1.21.0
scikit-learn==0.23.2
scipy == 1.5.0
soundfile==0.10.3.post1
pyworld == 0.2.10
PyYAML==5.4.1
tgt == 1.4.4
tqdm==4.60.0
unidecode == 1.1.1
mecab-python3
pytorch-lightning==1.7.2
num2words
transformers==4.12.3
pyopenjtalk